# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "base url image https://ubgmart.com/storage/ , s\u1eed d\u1ee5ng basic auth theo t\u00e0i kho\u1ea3n \u0111\u01b0\u1ee3c cung c\u1ea5p \u0111\u1ec3 truy c\u1eadp api",
        "termsOfService": "",
        "contact": {
            "email": "thanhan1507@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "api-ubg.test",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "\u0110\u0103ng k\u00fd th\u00e0nh vi\u00ean",
                "description": "Login user, s\u1eed d\u1ee5ng basic authentication b\u1eb1ng t\u00e0i kho\u1ea3n \u0111\u01b0\u1ee3c cung c\u1ea5p",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username as phone or email address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Input password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirm password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "\u0110\u0103ng nh\u1eadp",
                "description": "Login user, s\u1eed d\u1ee5ng basic authentication b\u1eb1ng t\u00e0i kho\u1ea3n \u0111\u01b0\u1ee3c cung c\u1ea5p",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username as phone or email address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Input password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "\u0110\u0103ng xu\u1ea5t",
                "description": "Logout user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "L\u1ea5y th\u00f4ng tin user",
                "description": "Return a user's information",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer Auth",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/api/auth/update-user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "C\u1eadp nh\u1eadt th\u00f4ng tin user",
                "description": "Update profile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Input name want to change",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Input email want to change",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Input phone want to change",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "description": "Input date of birth want to change",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "\u0110\u1ed5i m\u1eadt kh\u1ea9u",
                "description": "Reset password",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username as phone or email address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Input password want to change",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
...
}